vending machine using FPGA
This is the model of a vending machine that serves four products based on the rotation that is being assigned to it through the stepper motor upon authentication. this project can be improved by a card that reads information of the customers and deducts the price of the desired product.